Improve content scripts and prepare for build script
